Fix：詳細表示する登録データをDateFragmentでSQLiteから取得する
